backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
openai==1.30.1
pinecone==4.1.0
langchain==0.2.1
langchain-text-splitters==0.2.0
pdfplumber==0.11.0
python-dotenv==1.0.1
httpx==0.27.0